src/file.c: Proper EOF handling in seek
This change applies only to writable files.

The starting point `FILE_END` references the end of the file.  As more
data is written, this position increases.  However, the behavior
presented by StormLib is different.  The ultimate size of a file is
specified upon creation, thus `FILE_END` is constant (similar to that of
a read-only file), regardless of how much has been written.

We align the behavior surrounding `FILE_END` to be closer to that of a
file that increases in size.  It will always return the last position
written, rather than the file size (for which, a function already
exists).  This should align the behavior closer to that of Lua's I/O
library.