src/file.c: Support setting the buffering mode
StormLib buffers on writes by default.  However, we take a different
route and disable buffering by default.  We accomplish this by flushing
to disk after every write.  Note that this is an archive wide flush, so
it is obviously not very performant.  But it seems to be a sane default
in regards to archive integrity.

However, should one want to improve performance, `file:setvbuf ()` can
be used to enable buffering. In short, our full buffering mode
translates to us getting out of StormLib's way, for better or worse.