src: Remove support for flushing
For the most part this reverts the following commits:

- 5749a177c4329e5581590eac79b23ee9ee772c63.
- 2bd81d5db66e8c597cb4092576872af6a8425d6c.

The function `SFileFlushArchive ()` was being used for multiple
purposes.  First, it was being used to provide methods that allowed
explicitly flushing the archive and files.  Second, it was utilized to
ensure that the number of files was accurate (as it could go off while
data was in memory).  Third, it was believed it could be leveraged to
provide a no buffering mode.

It was discovered that `SFileFlushArchive ()` was not compatible with
`SFileWriteFile ()`.  At least, in a few key situations that very likely
could be encountered by the average user (and could not be prevented).
In short, if a flush occurred at the right point before an open writable
file was finished, the archive could be corrupted.  This is almost
certainly a bug in StormLib.  Sadly, I don't have the time to hunt this
one down, so we're dodging the issue basically.  However, it shuts down
all three previously mentioned cases.

In short, we cannot rely on the function for too much.  And its use
certainly needs to be extensively tested.  It should be noted that it is
sitll used, but in a very limited scope.  It is used before retrieving
the file count when checking if the archive limit needs to be increased.
This seems to present no problems.

In light of the removal of flush support, we go ahead and remove the
ability to get an archive's file count and limit.  There really isn't
too much need.  The archive will automatically increase its own limit
when adding or creating a file.  And if a user really wants the count,
then can use `mpq:list ()`.