src/mpq.c: Automatically close open file finders
It was possible for an open file finder to be garbage collected after an
associated MPQ archive.  This would lead to a segmentation fault.
However, unlike the related issue with open write enabled files, this
one cannot corrupt the MPQ archive.

To resolve this, we add file finders to the list of associated handle
objects that need to be garbage collected before the MPQ archive is
closed.