src/init.c: Allow creation of archives
Issue: #3

A new archive can be created via a `w+` flag passed to `stormlib.open
()`.  This will return a fresh archive to the user either by creating a
new one if one does not exist, or by truncating an existing file at the
specified path.

Note that the `w` flag is not supported, as StormLib does not provide
this functionality for an entire archive.  Not that it would be too
useful.