file.c: Actually fix `file:read (bytes)`
Issue: #6

This reverts: 70907141b026cba10001922df2dc907d133d50d7.

The 'fix' applied in that commit failed to realize the purpose of the
raw length check, which was to determine if a string with a length of
zero was the result from reading the file.  If so, it would return
`nil`, indcating end-of-file..  Removal of this check effectively broke
things further.

Diving into the issue further shows that Lua compat-5.3 uses
`lua_objlen` in Lua 5.1 to present `lua_rawlen`.  However, this
introduces behavior with numbers that is not present in Lua 5.2+:

```lua
    case LUA_TNUMBER: {
      size_t l;
      lua_lock(L);  /* `luaV_tostring' may create a new string */
      l = (luaV_tostring(L, o) ? tsvalue(o)->len : 0);
      lua_unlock(L);
      return l;
    }
```
Rather than address `lua_rawlen` being inconsistent across Lua versions,
we simply handle the end-of-file case and ensure that an empty string is
pushed onto the stack.  This ensures that the check works against a
string in all cases.