lua-stormlib: Provide 'custom' StormLib.h
Due to having only been building against StormLib 9.22, I had not run
into an issue with its C API that was introduced later.  A commit on
master introduced a default parameter, which is not allowed in C.  It
sounds as though upstream may prefer to not fix this breaking change to
the API, to avoid any breakage it may cause people using a broken API.

Anyways, introduced is a custom header with the default parameter
removed.  So long as a user uses StormLib >= 9.22 this appears to work
fine.  I did not test earlier versions.