file.c: Adjust seek behavior for writable files
The existing behavior was too complicated.  Adjusting the file position
for a writable file does not seem to do anything, except perhaps mess up
the finishing process.  As such, functionality has been reduced for such
files.  The use of `cur` will return the last written position, and
`end` will work as expected, and return the end of the file (i.e. file
size).